Удаление автором вопросов и ответов разнесено в разные спеки
